fix: sketch to fix a circular reference in dispatch
When reusing an existing model in another model that includes
additional models (aka. extending it) there is a circular reference
detected: `Type of property 'base' circularly references itself in
mapped type 'ExtractRematchDispatchersFromModels<ExtendedModel>'.`

The new test introduced will fail at the moment which is intended,
but a solution in `types.ts` is provided but the implications of the
change are not 100% clear to me.